dlog driver refactoring
