--- token/core/zkatdlog/nogh/v1/driver/ws_factory.go
@@ -1,61 +1,0 @@
-/*
-Copyright IBM Corp. All Rights Reserved.
-
-SPDX-License-Identifier: Apache-2.0
-*/
-
-package driver
-
-import (
-	"github.com/hyperledger-labs/fabric-smart-client/pkg/utils/errors"
-	"github.com/hyperledger-labs/fabric-smart-client/platform/view/services/metrics/disabled"
-	"github.com/hyperledger-labs/fabric-token-sdk/token/core"
-	v1 "github.com/hyperledger-labs/fabric-token-sdk/token/core/zkatdlog/nogh/v1/setup"
-	"github.com/hyperledger-labs/fabric-token-sdk/token/driver"
-	"github.com/hyperledger-labs/fabric-token-sdk/token/services/identity"
-	"github.com/hyperledger-labs/fabric-token-sdk/token/services/identity/membership"
-	"github.com/hyperledger-labs/fabric-token-sdk/token/services/logging"
-)
-
-// WalletServiceFactory is a factory for creating zkatdlog wallet services.
-type WalletServiceFactory struct {
-	*TokenDriverBase
-
-	storageProvider identity.StorageProvider
-}
-
-// NewWalletServiceFactory returns a new factory for the zkatdlog wallet service.
-func NewWalletServiceFactory(storageProvider identity.StorageProvider) core.NamedFactory[driver.WalletServiceFactory] {
-	return core.NamedFactory[driver.WalletServiceFactory]{
-		Name: core.DriverIdentifier(v1.DLogNoGHDriverName, v1.ProtocolV1),
-		Driver: &WalletServiceFactory{
-			TokenDriverBase: &TokenDriverBase{
-				Base: &Base{},
-			},
-			storageProvider: storageProvider},
-	}
-}
-
-// NewWalletService returns a new zkatdlog wallet service for the passed configuration and public parameters.
-func (d *WalletServiceFactory) NewWalletService(tmsConfig driver.Configuration, params driver.PublicParameters) (driver.WalletService, error) {
-	tmsID := tmsConfig.ID()
-	logger := logging.DriverLogger("token-sdk.driver.zkatdlog", tmsID.Network, tmsID.Channel, tmsID.Namespace)
-
-	pp, ok := params.(*v1.PublicParams)
-	if !ok {
-		return nil, errors.Errorf("invalid public parameters type [%T]", params)
-	}
-
-	return d.TokenDriverBase.NewWalletService(
-		tmsConfig,
-		&membership.NoBinder{},
-		d.storageProvider,
-		nil,
-		logger,
-		nil,
-		nil,
-		pp,
-		true,
-		&disabled.Provider{},
-	)
-}
